updates on review
